Updated cars.js
Updated functionality of the inheritance so that when a car have to be printed it checks if the car which has to inherit from also has to inherit properties. I made that as I made the inherit property non-enumerable and with the help of closure I managed to save the parent object whose properties I need for the inheritance.